feat(e2e): some more sonar complaints
